Index Autoconfiguration, overcoming filedescriptor limitations (#79)
This change implements auto-configuration of index settings, including configuring the maximum number of segments which is directly connected to the number of file descriptors. If not externally given it tries to increase the limit (usually from 1024 to 1M). In case of running low on file descriptors index operations are automatically throttled so that concurrent compaction can catch up and reduce open descriptors.

Also changes the queue implementation - after benchmarking - to use condition variables instead of a spin-lock if the queue runs full or empty. This improves indexing performance and also fixes a problem in the spin-lock implementation that could cause cpu hogging.