fix a corner case, reading a transition value which is spread across 2 chunks (#90)
fix a bug in reading transition values during construction which can be spread across 2 chunks.

This issue found in the code has probably not much impact in practice, nevertheless worth to fix. If a 
transition value is stored between 2 chunks (1st part in chunk x, 2nd part in chunk x+1) the value wasn't 
read correctly (chunking is only done at construction time, so lookup isn't affected). This read is 
part of minimization, so worst case this bug prevented minimization of this particular state.