Fix memory limit calculation for memory limits > 99272 MB (#206)
For memory limits > 99272 MB a bug in memory calculation triggered, resulting in super slow execution and loss of minimization. This change fixes the bug, ensuring the maximum possible configuration is returned and not configuration with 0. It also allows creating an even bigger hashtable by adding another prime. LBNL the minimum memory limit is checked and an exception is thrown.

fixes #205