ZipStateTraverser: support state traversal with inner weights (#217)
This change add support for weighted state traversal in comparable/zip state traverser. This is the 1st step in order to support completions with inner weights from keyvi indexes.

When combining the traversers, inner weights have to be normalized, that means a global weight map has to be created and transitions need to be sorted the same way for all inner traversers. This can be expensive, however as segments get merged quite quickly, this is probably only a problem for the 1st levels in the tree.