portable c-char in dictionary.rs
`char` type is not specifically defined. It can be signed or unsigned depends
on architecture.

On x86 `char` is signed but on some ARM platforms char can be signed or unsigned.

```

int main() {
  char c = -1;

  if (c > 0)
      printf("c is positive \n");  // Docker container on Apple M1
  else
      printf("c is negative \n");  // Native Apple M1
}

```

The output will be "c is negative" on Apple M1 and "c is positibe" on linux
docker container on Apple M1.

Let's use `c_char` which is portable.

More details:
https://developer.arm.com/documentation/den0013/d/Porting/Miscellaneous-C-porting-issues/unsigned-char-and-signed-char
https://doc.rust-lang.org/std/os/raw/type.c_char.html